@@ -25,6 +25,7 @@
     "@fortawesome/free-solid-svg-icons": "^6.7.1",
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@mapbox/mapbox-gl-draw": "^1.4.3",
+    "@testing-library/jest-dom": "^6.6.3",
     "@turf/turf": "^7.1.0",
     "bcryptjs": "^2.4.3",
     "chaire-lib-common": "^0.2.2",
@@ -42,11 +43,11 @@
     "moment": "^2.30.1",
     "moment-business-days": "^1.2.0",
     "papaparse": "^5.4.1",
-    "react": "^16.14.0",
+    "react": "^19.0.0",
     "react-collapsible": "^2.10.0",
     "react-color": "^2.19.3",
     "react-datepicker": "^3.8.0",
-    "react-dom": "^16.14.0",
+    "react-dom": "^19.0.0",
     "react-i18next": "^12.3.1",
     "react-loadable": "^5.5.0",
     "react-markdown": "^8.0.7",
@@ -64,9 +65,7 @@
     "typescript": "^5.7.2"
   },
   "devDependencies": {
-    "@testing-library/react": "^10.4.9",
-    "@types/enzyme": "^3.10.18",
-    "@types/enzyme-adapter-react-16": "^1.0.9",
+    "@testing-library/react": "^16.1.0",
     "@types/geojson": "^7946.0.15",
     "@types/jest": "^29.5.14",
     "@types/lodash": "^4.17.13",
@@ -77,15 +76,11 @@
     "@types/react-redux": "^7.1.9",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-select": "^5.0.1",
-    "@types/react-test-renderer": "^16.9.12",
     "@typescript-eslint/eslint-plugin": "^8.17.0",
     "@typescript-eslint/parser": "^8.17.0",
     "copyfiles": "^2.4.1",
     "cross-env": "^7.0.3",
     "dotenv": "^16.4.7",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.8",
-    "enzyme-to-json": "^3.6.2",
     "eslint": "^8.57.1",
     "eslint-plugin-n": "^15.7.0",
     "eslint-plugin-react": "^7.37.2",
@@ -95,7 +90,6 @@
     "mockdate": "^3.0.5",
     "prettier-eslint-cli": "^8.0.1",
     "react-select-event": "^5.5.1",
-    "react-test-renderer": "^16.14.0",
     "redux-mock-store": "^1.5.5",
     "rimraf": "^6.0.1",
     "ts-jest": "^29.2.5"
