@@ -28,6 +28,7 @@
     "@fortawesome/react-fontawesome": "^0.2.2",
     "@mapbox/mapbox-gl-draw": "^1.4.3",
     "@turf/turf": "^7.1.0",
+    "@types/react-dom": "^19.0.2",
     "bcryptjs": "^2.4.3",
     "chaire-lib-common": "^0.2.2",
     "chaire-lib-frontend": "^0.2.2",
@@ -42,11 +43,11 @@
     "moment-business-days": "^1.2.0",
     "papaparse": "^5.4.1",
     "rc-menu": "^9.16.0",
-    "react": "^16.14.0",
+    "react": "^19.0.0",
     "react-collapsible": "^2.10.0",
     "react-color": "^2.19.3",
     "react-datepicker": "^3.8.0",
-    "react-dom": "^16.14.0",
+    "react-dom": "^19.0.0",
     "react-i18next": "^12.3.1",
     "react-mathjax": "^1.0.1",
     "react-redux": "^7.2.9",
@@ -67,17 +68,16 @@
   },
   "devDependencies": {
     "@alienfast/i18next-loader": "^2.0.2",
-    "@testing-library/react": "^10.4.9",
+    "@testing-library/react": "^16.1.0",
     "@types/geojson": "^7946.0.15",
     "@types/jest": "^29.5.14",
     "@types/lodash": "^4.17.13",
     "@types/mapbox-gl": "^2.3.1",
-    "@types/osrm": "^5.25.5",
     "@types/node": "^22.10.1",
+    "@types/osrm": "^5.25.5",
     "@types/react-mathjax": "^1.0.4",
     "@types/react-redux": "^7.1.9",
     "@types/react-router-dom": "^5.3.3",
-    "@types/react-test-renderer": "^16.9.12",
     "@types/socket.io-client": "^1.4.36",
     "@typescript-eslint/eslint-plugin": "^8.17.0",
     "@typescript-eslint/parser": "^8.17.0",
@@ -88,8 +88,6 @@
     "cross-env": "^7.0.3",
     "css-loader": "^6.11.0",
     "dotenv": "^16.4.7",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.8",
     "eslint": "^8.57.1",
     "eslint-plugin-n": "^15.7.0",
     "eslint-plugin-react": "^7.37.2",
@@ -100,7 +98,6 @@
     "mini-css-extract-plugin": "^2.9.2",
     "mockdate": "^3.0.5",
     "prettier-eslint-cli": "^8.0.1",
-    "react-test-renderer": "^16.14.0",
     "redux-mock-store": "^1.5.5",
     "rimraf": "^6.0.1",
     "sass": "^1.82.0",
