@@ -5,56 +5,55 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 import * as React from 'react';
-import { create } from 'react-test-renderer';
+import { fireEvent, render } from '@testing-library/react';
+import '@testing-library/jest-dom';
+
 import InputColor from '../InputColor';
-import { shallow } from 'enzyme';
 
 const mockOnChange = jest.fn();
 const testId = 'ColorWidgetId';
 
 test('Default props', () => {
-    const input = create(<InputColor
+    const { container } = render(<InputColor
         id = {testId}
         onValueChange = {mockOnChange}
         defaultColor = '#554433'
-    />)
-        .toJSON();
-    expect(input).toMatchSnapshot();
+    />);
+    expect(container.firstChild).toMatchSnapshot();
 });
 
 test('All props', () => {
-    const input = create(<InputColor
+    const { container } = render(<InputColor
         id = {testId}
         onValueChange = {mockOnChange}
         value = '#123456'
         defaultColor = '#123456'
-    />)
-        .toJSON();
-    expect(input).toMatchSnapshot();
+    />);
+    expect(container.firstChild).toMatchSnapshot();
 });
 
 test('Blank value', () => {
-    const input = create(<InputColor
+    const { container } = render(<InputColor
         id = {testId}
         onValueChange = {mockOnChange}
         value = ''
         defaultColor = '#223344'
-    />)
-        .toJSON();
-    expect(input).toMatchSnapshot();
+    />);
+    expect(container.firstChild).toMatchSnapshot();
 });
 
 test('With colorpicker', () => {
-    const input = shallow(<InputColor
+    const { container } = render(<InputColor
         id = {testId}
         onValueChange = {mockOnChange}
         value = '#123456'
         defaultColor = '#223344'
     />);
-    input.setState({displayColorPicker: true});
-    expect(input.html()).toMatchSnapshot();
+    const colorPickerOpenButton = container.querySelector('._open-color-picker') as HTMLInputElement;
+    fireEvent.click(colorPickerOpenButton);
+    expect(container.firstChild).toMatchSnapshot();
 
-    input.setState({displayColorPicker: false});
-    expect(input.html()).toMatchSnapshot();
+    const colorPickerCloseButton = container.querySelector('._close-color-picker') as HTMLInputElement;
+    fireEvent.click(colorPickerCloseButton);
+    expect(container.firstChild).toMatchSnapshot();
 });
-
