@@ -5,37 +5,37 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 import * as React from 'react';
-import { create } from 'react-test-renderer';
+import { render, screen, fireEvent } from '@testing-library/react';
+import '@testing-library/jest-dom';
+
 import InputSelect from '../InputSelect';
-import { render, fireEvent } from "@testing-library/react";
 
 const mockOnChange = jest.fn();
-const testId = "SelectWidgetId";
-const testLabel = "Select label";
-const defaultChoiceValue = "test1";
-const anotherChoiceValue = "test2";
+const testId = 'SelectWidgetId';
+const testLabel = 'Select label';
+const defaultChoiceValue = 'test1';
+const anotherChoiceValue = 'test2';
 const testChoices = [
     { value: defaultChoiceValue },
-    { value: anotherChoiceValue, label: "label for test2" },
-    { value: "disabled", disabled: true},
-    { value: "has sub-choices",
+    { value: anotherChoiceValue, label: 'label for test2' },
+    { value: 'disabled', disabled: true },
+    { value: 'has sub-choices',
         choices: [
-            { value: "subChoice1" },
-            { value: "subChoice2", label: "label for subchoice2" }
+            { value: 'subChoice1' },
+            { value: 'subChoice2', label: 'label for subchoice2' }
         ]
     }
 ];
 
 test('Default props', () => {
-    const input = create(<InputSelect
+    const input = render(<InputSelect
         id = {testId}
-    />)
-        .toJSON();
+    />);
     expect(input).toMatchSnapshot();
 });
 
 test('All props', () => {
-    const input = create(<InputSelect
+    const input = render(<InputSelect
         id = {testId}
         onValueChange = {mockOnChange}
         value = {anotherChoiceValue}
@@ -44,45 +44,42 @@ test('All props', () => {
         noBlank = {true}
         choices = {testChoices}
         localePrefix = "something"
-        t = {str => str}
-    />)
-        .toJSON();
+        t = {(str) => str}
+    />);
     expect(input).toMatchSnapshot();
 });
 
 test('Disabled', () => {
-    const input = create(<InputSelect
+    const input = render(<InputSelect
         id = {testId}
         onValueChange = {mockOnChange}
         disabled = {true}
-    />)
-        .toJSON();
+    />);
     expect(input).toMatchSnapshot();
 });
 
 test('With blank choice', () => {
-    const input = create(<InputSelect
+    const input = render(<InputSelect
         id = {testId}
         onValueChange = {mockOnChange}
         noBlank = {false}
         choices = {testChoices}
-    />)
-        .toJSON();
+    />);
     expect(input).toMatchSnapshot();
 });
 
 test('Default value', () => {
     mockOnChange.mockClear();
     const { getByLabelText } = render(
-    <div>
-        <InputSelect
-            id = {testId}
-            onValueChange = {mockOnChange}
-            defaultValue = {defaultChoiceValue}
-            choices = {testChoices}
-        />
-        <label htmlFor={testId}>{testLabel}</label>
-    </div>);
+        <div>
+            <InputSelect
+                id = {testId}
+                onValueChange = {mockOnChange}
+                defaultValue = {defaultChoiceValue}
+                choices = {testChoices}
+            />
+            <label htmlFor={testId}>{testLabel}</label>
+        </div>);
     const input = getByLabelText(testLabel) as HTMLInputElement;
     expect(input.value).toBe(defaultChoiceValue);
 });
@@ -90,34 +87,34 @@ test('Default value', () => {
 test('Default and initial value', () => {
     mockOnChange.mockClear();
     const { getByLabelText } = render(
-    <div>
-        <InputSelect
-            id = {testId}
-            onValueChange = {mockOnChange}
-            defaultValue = {defaultChoiceValue}
-            value = {anotherChoiceValue}
-            choices = {testChoices}
-        />
-        <label htmlFor={testId}>{testLabel}</label>
-    </div>);
+        <div>
+            <InputSelect
+                id = {testId}
+                onValueChange = {mockOnChange}
+                defaultValue = {defaultChoiceValue}
+                value = {anotherChoiceValue}
+                choices = {testChoices}
+            />
+            <label htmlFor={testId}>{testLabel}</label>
+        </div>);
     const input = getByLabelText(testLabel) as HTMLInputElement;
     expect(input.value).toBe(anotherChoiceValue);
 });
 
 test('Default and initial empty value', () => {
     mockOnChange.mockClear();
-    const value = "";
+    const value = '';
     const { getByLabelText } = render(
-    <div>
-        <InputSelect
-            id = {testId}
-            onValueChange = {mockOnChange}
-            defaultValue = {defaultChoiceValue}
-            value = {value}
-            choices = {testChoices}
-        />
-        <label htmlFor={testId}>{testLabel}</label>
-    </div>);
+        <div>
+            <InputSelect
+                id = {testId}
+                onValueChange = {mockOnChange}
+                defaultValue = {defaultChoiceValue}
+                value = {value}
+                choices = {testChoices}
+            />
+            <label htmlFor={testId}>{testLabel}</label>
+        </div>);
     const input = getByLabelText(testLabel) as HTMLInputElement;
     expect(input.value).toBe(value);
 });
@@ -125,17 +122,17 @@ test('Default and initial empty value', () => {
 test('Invalid default and value', () => {
     mockOnChange.mockClear();
     const { getByLabelText } = render(
-    <div>
-        <InputSelect
-            id = {testId}
-            onValueChange = {mockOnChange}
-            defaultValue = "not a value"
-            value = "still not a choice"
-            choices = {testChoices}
-            noBlank = {true}
-        />
-        <label htmlFor={testId}>{testLabel}</label>
-    </div>);
+        <div>
+            <InputSelect
+                id = {testId}
+                onValueChange = {mockOnChange}
+                defaultValue = "not a value"
+                value = "still not a choice"
+                choices = {testChoices}
+                noBlank = {true}
+            />
+            <label htmlFor={testId}>{testLabel}</label>
+        </div>);
     const input = getByLabelText(testLabel) as HTMLInputElement;
     expect(input.value).toBe(defaultChoiceValue);
 });
@@ -143,16 +140,16 @@ test('Invalid default and value', () => {
 test('Call onChange', () => {
     mockOnChange.mockClear();
     const { getByLabelText } = render(
-    <div>
-        <InputSelect
-            id = {testId}
-            onValueChange = {mockOnChange}
-            choices = {testChoices}
-        />
-        <label htmlFor={testId}>{testLabel}</label>
-    </div>);
+        <div>
+            <InputSelect
+                id = {testId}
+                onValueChange = {mockOnChange}
+                choices = {testChoices}
+            />
+            <label htmlFor={testId}>{testLabel}</label>
+        </div>);
     const input = getByLabelText(testLabel) as HTMLInputElement;
-    const newText = "new text";
-    fireEvent.change(input, {target: { value: newText}});
+    const newText = 'new text';
+    fireEvent.change(input, { target: { value: newText } });
     expect(mockOnChange).toHaveBeenCalledTimes(1);
 });
