@@ -5,24 +5,25 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 import * as React from 'react';
-import { create } from 'react-test-renderer';
+import { render, screen, fireEvent } from '@testing-library/react';
+import '@testing-library/jest-dom';
+
 import InputText from '../InputText';
-import { render, fireEvent } from "@testing-library/react";
+import { create } from 'node:domain';
 
 const mockOnChange = jest.fn();
-const testId = "textAreaWidgetId";
-const testLabel = "Text Area";
+const testId = 'textAreaWidgetId';
+const testLabel = 'Text Area';
 
 test('Default props', () => {
-    const input = create(<InputText
+    const input = render(<InputText
         id = {testId}
-    />)
-        .toJSON();
+    />);
     expect(input).toMatchSnapshot();
 });
 
 test('All props', () => {
-    const input = create(<InputText
+    const input = render(<InputText
         id = {testId}
         onValueChange = {mockOnChange}
         rows = {10}
@@ -30,46 +31,43 @@ test('All props', () => {
         maxLength = {100}
         disabled = {false}
         placeholder = "placeholder"
-    />)
-        .toJSON();
+    />);
     expect(input).toMatchSnapshot();
 });
 
 test('Disabled', () => {
-    const input = create(<InputText
+    const input = render(<InputText
         id = {testId}
         onValueChange = {mockOnChange}
         disabled = {true}
-    />)
-        .toJSON();
+    />);
     expect(input).toMatchSnapshot();
 });
 
 test('Placeholder', () => {
-    const input = create(<InputText
+    const input = render(<InputText
         id = {testId}
         onValueChange = {mockOnChange}
         rows = {10}
         maxLength = {100}
         disabled = {false}
         placeholder = "placeholder"
-    />)
-        .toJSON();
+    />);
     expect(input).toMatchSnapshot();
 });
 
 test('Call onChange', () => {
     mockOnChange.mockClear();
     const { getByLabelText } = render(
-    <div>
-        <InputText
-            id = {testId}
-            onValueChange = {mockOnChange}
-        />
-        <label htmlFor={testId}>{testLabel}</label>
-    </div>);
+        <div>
+            <InputText
+                id = {testId}
+                onValueChange = {mockOnChange}
+            />
+            <label htmlFor={testId}>{testLabel}</label>
+        </div>);
     const input = getByLabelText(testLabel) as HTMLInputElement;
-    const newText = "new text";
-    fireEvent.change(input, {target: { value: newText}});
+    const newText = 'new text';
+    fireEvent.change(input, { target: { value: newText } });
     expect(mockOnChange).toHaveBeenCalledTimes(1);
 });
