@@ -5,12 +5,12 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 import * as React from 'react';
-import { create } from 'react-test-renderer';
-import { mount } from 'enzyme';
 
 import SelectedObjectButtons from '../SelectedObjectButtons';
 import { ObjectWithHistory } from 'chaire-lib-common/lib/utils/objects/ObjectWithHistory';
 import { GenericAttributes } from 'chaire-lib-common/lib/utils/objects/GenericObject';
+import { render, screen, fireEvent } from '@testing-library/react';
+import '@testing-library/jest-dom';
 
 // TODO Cannot do a stub class, it gives compilation error. Wait for a mocking/stubbing library
 let newObject: ObjectWithHistory<GenericAttributes>;
@@ -23,130 +23,115 @@ beforeEach(() => {
     existingObject.startEditing();
     newObject = new ObjectWithHistory({}, true);
     newObject.startEditing();
-})
+});
 
 test('Test with delete button', () => {
-    const buttons = create(<SelectedObjectButtons
+    const { container } = render(<SelectedObjectButtons
         object = {existingObject}
-    />)
-        .toJSON();
-    expect(buttons).toMatchSnapshot();
+    />);
+    expect(container).toMatchSnapshot();
 });
 
 test('Test without delete button', () => {
-    const buttons = create(<SelectedObjectButtons
+    const { container } = render(<SelectedObjectButtons
         object = {existingObject}
         hideDelete = {true}
-    />)
-        .toJSON();
-    expect(buttons).toMatchSnapshot();
+    />);
+    expect(container).toMatchSnapshot();
 });
 
 test('Test with save button', () => {
-    const buttons = create(<SelectedObjectButtons
+    const { container } = render(<SelectedObjectButtons
         object = {existingObject}
-    />)
-        .toJSON();
-    expect(buttons).toMatchSnapshot();
+    />);
+    expect(container).toMatchSnapshot();
 });
 
 test('Test without save button', () => {
-    const buttons = create(<SelectedObjectButtons
+    const { container } = render(<SelectedObjectButtons
         object = {existingObject}
         hideSave = {true}
-    />)
-        .toJSON();
-    expect(buttons).toMatchSnapshot();
+    />);
+    expect(container).toMatchSnapshot();
 });
 
 test('Test back click on unmodified object', () => {
     const mockBackAction = jest.fn();
     const mockOpenModal = jest.fn();
-    const buttons = mount(<SelectedObjectButtons
+    render(<SelectedObjectButtons
         object = {existingObject}
         backAction = {mockBackAction}
         openBackConfirmModal = {mockOpenModal}
     />);
 
-    const backButtonWrapper = buttons.find({ type: 'button' }).at(0);
-    const backButton = buttons.find({ type: 'button' }).at(1);
-    expect(backButtonWrapper.key()).toEqual('back');
-    backButton.simulate('click');
+    const backButton = screen.getByTitle('main:Back');
+    fireEvent.click(backButton);
 
     expect(mockBackAction).toHaveBeenCalledTimes(1);
     expect(mockOpenModal).toHaveBeenCalledTimes(0);
-})
+});
 
 test('Test back click on modified object', () => {
     const mockBackAction = jest.fn();
     const mockOpenModal = jest.fn();
-    const buttons = mount(<SelectedObjectButtons
-        object = {existingObject}
-        backAction = {mockBackAction}
-        openBackConfirmModal = {mockOpenModal}
+
+    existingObject.setData('modification', 1234);
+    /*
+    With enzyme, we can change the object after render, but for a weird reason,
+    with testing-library, we cannot. maybe the rendering just clones the atributes.
+    So we need to do the change before the render.
+    */
+    render(<SelectedObjectButtons
+        object={existingObject}
+        backAction={mockBackAction}
+        openBackConfirmModal={mockOpenModal}
     />);
 
-    const backButtonWrapper = buttons.find({ type: 'button' }).at(0);
-    const backButton = buttons.find({ type: 'button' }).at(1);
-    expect(backButtonWrapper.key()).toEqual('back');
-
-    // Do a change on the object and set the props again
-    const dataToModify = 'modification';
-    const value = 1234;
-    existingObject.setData(dataToModify, value);
-    buttons.setProps({
-        object: existingObject,
-        onBack: mockBackAction,
-        openBackConfirmModal: mockOpenModal,
-    });
-
-    backButton.simulate('click');
+    const backButton = screen.getByTitle('main:Back');
+    fireEvent.click(backButton);
+    console.log('hasCHanged');
 
     expect(mockOpenModal).toHaveBeenCalledTimes(1);
     expect(mockBackAction).toHaveBeenCalledTimes(0);
-})
+});
 
 test('Test undo click', () => {
     const mockUndo = jest.fn();
-    const buttons = mount(<SelectedObjectButtons
+    const { rerender } = render(<SelectedObjectButtons
         object = {existingObject}
         onUndo = {mockUndo}
     />);
 
     // Should be disabled now
-    const undoButtonWrapper = buttons.find({ type: 'button' }).at(2);
-    const undoButton = buttons.find({ type: 'button' }).at(3);
-    expect(undoButtonWrapper.key()).toEqual('undo');
-    expect(undoButton.getDOMNode<HTMLInputElement>().disabled).toBe(true);
+    const undoButton = screen.getByTitle('main:Undo');
+    expect(undoButton).toBeDisabled();
 
     // Do a change on the object and set the props again
     const dataToUndo = 'toUndo';
     const value = 1234;
     existingObject.startEditing();
     existingObject.setData(dataToUndo, value);
-    buttons.setProps({
-        object: existingObject,
-        onUndo: mockUndo
-    });
-    expect(undoButton.getDOMNode<HTMLInputElement>().disabled).toBe(false);
-    undoButton.simulate('click');
+    rerender(<SelectedObjectButtons
+        object={existingObject}
+        onUndo={mockUndo}
+    />);
+    expect(undoButton).not.toBeDisabled();
+    fireEvent.click(undoButton);
     expect(mockUndo).toHaveBeenCalledTimes(1);
     expect(existingObject.getData(dataToUndo)).toBeUndefined();
     existingObject.stopEditing();
-})
+});
 
 test('Test redo click', () => {
     const mockRedo = jest.fn();
-    const buttons = mount(<SelectedObjectButtons
+    const { rerender } = render(<SelectedObjectButtons
         object = {existingObject}
         onRedo = {mockRedo}
     />);
 
     // Should be disabled now
-    const redoButtonWrapper = buttons.find({ type: 'button' }).at(4);
-    const redoButton = buttons.find({ type: 'button' }).at(5);
-    expect(redoButtonWrapper.key()).toEqual('redo');
-    expect(redoButton.getDOMNode<HTMLInputElement>().disabled).toBe(true);
+    const redoButton = screen.getByTitle('main:Redo');
+    expect(redoButton).toBeDisabled();
 
     // Do a change on the object and set the props again
     const dataToUndo = 'toUndo';
@@ -154,60 +139,56 @@ test('Test redo click', () => {
     existingObject.startEditing();
     existingObject.setData(dataToUndo, value);
     existingObject.undo();
-    buttons.setProps({
-        object: existingObject,
-        onRedo: mockRedo
-    });
-    expect(redoButton.getDOMNode<HTMLInputElement>().disabled).toBe(false);
-    redoButton.simulate('click');
+    rerender(<SelectedObjectButtons
+        object={existingObject}
+        onRedo={mockRedo}
+    />);
+    expect(redoButton).not.toBeDisabled();
+    fireEvent.click(redoButton);
     expect(mockRedo).toHaveBeenCalledTimes(1);
     expect(existingObject.getData(dataToUndo)).toEqual(value);
     existingObject.stopEditing();
-})
+});
 
 test('Test default save click', () => {
 
     // TODO Wait for stubbing library for easier test
-})
+});
 
 test('Test custom save action', () => {
 
     // TODO Wait for stubbing library for easier test
-})
+});
 
 test('Test delete for new objects', () => {
     const mockDeleteAction = jest.fn();
     const mockOpenModal = jest.fn();
-    const buttons = mount(<SelectedObjectButtons
-        object = {newObject}
-        onDelete = {mockDeleteAction}
-        openDeleteConfirmModal = {mockOpenModal}
+    render(<SelectedObjectButtons
+        object={newObject}
+        onDelete={mockDeleteAction}
+        openDeleteConfirmModal={mockOpenModal}
     />);
 
-    const deleteButtonWrapper = buttons.find({ type: 'button' }).at(8);
-    const deleteButton = buttons.find({ type: 'button' }).at(9);
-    expect(deleteButtonWrapper.key()).toEqual('delete');
-    deleteButton.simulate('click');
+    const deleteButton = screen.getByTitle('main:Delete');
+    fireEvent.click(deleteButton);
 
     expect(mockDeleteAction).toHaveBeenCalledTimes(1);
     expect(mockOpenModal).toHaveBeenCalledTimes(0);
-})
+});
 
 test('Test delete for existing objects', () => {
     const mockDeleteAction = jest.fn();
     const mockOpenModal = jest.fn();
-    const buttons = mount(<SelectedObjectButtons
-        object = {existingObject}
-        onDelete = {mockDeleteAction}
-        openDeleteConfirmModal = {mockOpenModal}
+
+    render(<SelectedObjectButtons
+        object={existingObject}
+        onDelete={mockDeleteAction}
+        openDeleteConfirmModal={mockOpenModal}
     />);
 
-    const deleteButtonWrapper = buttons.find({ type: 'button' }).at(8);
-    const deleteButton = buttons.find({ type: 'button' }).at(9);
-    expect(deleteButtonWrapper.key()).toEqual('delete');
-    deleteButton.simulate('click');
+    const deleteButton = screen.getByTitle('main:Delete');
+    fireEvent.click(deleteButton);
 
     expect(mockOpenModal).toHaveBeenCalledTimes(1);
     expect(mockDeleteAction).toHaveBeenCalledTimes(0);
-})
-
+});
