@@ -5,9 +5,11 @@
  * License text available at https://opensource.org/licenses/MIT
  */
 import * as React from 'react';
-import { create } from 'react-test-renderer';
+import { render, screen, fireEvent } from '@testing-library/react';
+import '@testing-library/jest-dom';
+
 import ButtonCell from '../ButtonCell';
-import { render, fireEvent } from "@testing-library/react";
+import { create } from 'node:domain';
 
 // Mock react-markdown and remark-gfm as they use syntax not supported by jest
 jest.mock('react-markdown', () => 'Markdown');
@@ -16,40 +18,36 @@ jest.mock('remark-gfm', () => 'remark-gfm');
 const mockOnClick = jest.fn();
 
 test('Left align, one child', () => {
-    const input = create(<ButtonCell
+    const input = render(<ButtonCell
         key = 'key'
         alignment = 'left'
-    >test</ButtonCell>)
-        .toJSON();
+    >test</ButtonCell>);
     expect(input).toMatchSnapshot();
 });
 
 test('Rigth align, one child', () => {
-    const input = create(<ButtonCell
+    const input = render(<ButtonCell
         key = 'key'
         alignment = 'right'
-    >test</ButtonCell>)
-        .toJSON();
+    >test</ButtonCell>);
     expect(input).toMatchSnapshot();
 });
 
 test('Flush align, one child', () => {
-    const input = create(<ButtonCell
+    const input = render(<ButtonCell
         key = 'key'
         alignment = 'flush'
-    >test</ButtonCell>)
-        .toJSON();
+    >test</ButtonCell>);
     expect(input).toMatchSnapshot();
 });
 
 test('More complex children, onClick', () => {
-    const input = create(<ButtonCell
+    const input = render(<ButtonCell
         key = 'key'
         alignment = 'left'
         onClick = {mockOnClick}
         title = 'title'
-    ><h2>test</h2>other text</ButtonCell>)
-        .toJSON();
+    ><h2>test</h2>other text</ButtonCell>);
     expect(input).toMatchSnapshot();
 });
 
